Display a progress view when the player is available yet
