--- Sources/Castor/UserInterface/Player/PlayerMainView.swift
@@ -1,103 +1,0 @@
-//
-//  Copyright (c) SRG SSR. All rights reserved.
-//
-//  License information is available from the LICENSE file.
-//
-
-import GoogleCast
-import SwiftUI
-
-private enum GeometryEffectIdentifier: Hashable {
-    case artwork
-    case info
-}
-
-struct PlayerMainView: View {
-    @ObservedObject var player: CastPlayer
-    let device: CastDevice?
-
-    @State private var isPlaylistPresented = false
-    @Namespace var namespace
-    @Environment(\.verticalSizeClass) private var verticalSizeClass
-
-    var body: some View {
-        if verticalSizeClass == .compact {
-            HStack(spacing: 0) {
-                VStack {
-                    HStack(spacing: 20) {
-                        visualView()
-                        informationView()
-                    }
-                    .frame(height: 100)
-                    Spacer()
-                    ControlsView(player: player, isPlaylistPresented: $isPlaylistPresented)
-                }
-                PlaylistView(player: player)
-            }
-        }
-        else {
-            VStack(spacing: 0) {
-                if isPlaylistPresented {
-                    HStack(spacing: 20) {
-                        visualView()
-                        informationView()
-                    }
-                    .frame(height: 100)
-                    .padding([.horizontal, .top])
-                    PlaylistView(player: player)
-                        .transition(.move(edge: .bottom))
-                }
-                else {
-                    Spacer()
-                    visualView()
-                        .padding()
-                    Spacer()
-                    informationView()
-                        .padding(.horizontal)
-                }
-                ControlsView(player: player, isPlaylistPresented: $isPlaylistPresented)
-            }
-            .animation(.default, value: isPlaylistPresented)
-        }
-    }
-
-    private func artworkImage() -> some View {
-        ArtworkImage(url: player.metadata?.imageUrl())
-    }
-
-    private func loadingIndicator() -> some View {
-        ProgressView()
-            .tint(.white)
-            .padding(10)
-            .background(
-                Circle()
-                    .fill(Color(white: 0, opacity: 0.4))
-            )
-            .opacity(player.isBusy ? 1 : 0)
-            .animation(.default, value: player.isBusy)
-    }
-
-    private func informationView() -> some View {
-        VStack(alignment: .leading) {
-            LiveLabel(player: player)
-            if let title = player.metadata?.title {
-                Text(title)
-            }
-            if let device {
-                Text("Connected to \(device.name ?? "receiver")")
-                    .font(.subheadline)
-                    .foregroundStyle(.secondary)
-            }
-        }
-        .frame(maxWidth: .infinity, alignment: .leading)
-        .matchedGeometryEffect(id: GeometryEffectIdentifier.info, in: namespace)
-    }
-
-    private func visualView() -> some View {
-        ZStack {
-            artworkImage()
-            loadingIndicator()
-        }
-        .matchedGeometryEffect(id: GeometryEffectIdentifier.artwork, in: namespace)
-    }
-}
